rename: firebase and sheet handlers
