+++ commands/find-user.js
@@ -1,5 +1,5 @@
 const Discord = require('discord.js');
-const { getMember } = require('../firebase/firebase_handler');
+const { getMember } = require('../firebase');
 const { PREFIX } = require('../utils/constants');
 const { sendDissapearingMessage } = require('../utils/functions');
