+++ commands/update-database.js
@@ -1,6 +1,6 @@
 const Discord = require('discord.js');
-const { updateDatabase } = require('../excel/spreadsheet_handler');
-const { getMember, getAllMember } = require('../firebase/firebase_handler');
+const { updateDatabase } = require('../excel');
+const { getMember, getAllMember } = require('../firebase');
 const { PREFIX, COLORS } = require('../utils/constants');
 const { sendDissapearingMessage } = require('../utils/functions');
