+++ events/guildMemberUpdate.js
@@ -1,6 +1,6 @@
 // eslint-disable-next-line no-unused-vars
 const Discord = require('discord.js');
-const { updateUser } = require('../firebase/firebase_handler');
+const { updateUser } = require('../firebase');
 
 module.exports = {
     name: 'guildMemberUpdate',
