+++ events/message.js
@@ -1,5 +1,5 @@
 const Discord = require('discord.js');
-const { firebaseSpamLinkArray } = require('../firebase/firebase_handler');
+const { firebaseSpamLinkArray } = require('../firebase');
 const { COLORS, PREFIX } = require('../utils/constants');
 const { logger } = require('../utils/logger');
