--- index.js
@@ -1,12 +1,0 @@
-const express = require('express');
-
-const app = express();
-
-app.use(express.json());
-app.use(express.urlencoded({ extended: false }));
-
-app.get('/', (req, res) => res.send('Master Oogway'));
-
-require('./app');
-
-module.exports = app;
